rework: rewrite PointLightAirlockSystem on client
