+++ Content.Client/_White/Lighting/PointLight/Airlock/PointLightAirlockComponent.cs
