+++ Content.Shared/_White/Lighting/PointLight/Airlock/PointLightAirlockComponent.cs
@@ -6,20 +6,20 @@ namespace Content.Shared._White.Lighting.PointLight.Airlock;
 public sealed partial class PointLightAirlockComponent : Component
 {
     [ViewVariables]
-    public bool IsPowered;
-
-    [ViewVariables]
     public bool IsLightsEnabled;
 
     [ViewVariables]
-    public string RedColor = "#D56C6C";
+    public bool LastBoltsState;
 
     [ViewVariables]
-    public string BlueColor = "#7F93C0";
+    public readonly string RedColor = "#D56C6C";
 
     [ViewVariables]
-    public string YellowColor = "#BDC07F";
+    public readonly string BlueColor = "#7F93C0";
 
     [ViewVariables]
-    public string GreenColor = "#7FC080";
+    public readonly string YellowColor = "#BDC07F";
+
+    [ViewVariables]
+    public readonly string GreenColor = "#7FC080";
 }
